출고 삭제 수정
root-context.xml파일에 트랜잭션 설정 추가
삭제 후 업데이트가 아니라 업데이트 후 삭제로 순서 변경

+++ src/main/java/com/project/ManagementProgram/InventoryController.java
@@ -260,10 +260,19 @@ public class InventoryController {
 	}
 	
 	@RequestMapping("/inventoryGoodsLend/remove")
+	@Transactional(rollbackFor = {RuntimeException.class, Exception.class})
 	public String goodsRemove(Goods goods, RedirectAttributes rattr){
+
 		try {
-			int res = inventoryservice.removeGoods(goods.getGno());
 			int gcount = inventoryservice.updateInventoryGoodsCount(goods);
+			
+			if(gcount < 1)
+				throw new Exception("Count update Error");
+			
+			int res = inventoryservice.removeGoods(goods.getGno());
+			
+			if(res < 1)
+				throw new Exception("Remove Error");
 			
 			rattr.addFlashAttribute("msg", "DEL_OK");
 		}catch(Exception e) {
